+++ examples/create-patient/main.go
@@ -14,8 +14,8 @@ import (
 func main() {
 	m, err := medplum.New(&medplum.Options{
 		MedplumURL:   "http://localhost:8103",
-		ClientID:     "3008218e-5de9-4398-a987-ca393e3e64b0",
-		ClientSecret: "1b6b7708423fa6cc589d2996e40d35bc2ba38d6af366e16660bcfcecb5438896",
+		ClientID:     "foo",
+		ClientSecret: "bar",
 	})
 
 	if err != nil {